useDeviceSize refactor
