+++ src/hooks/index.ts
@@ -1,6 +1,5 @@
 export * from './useApi';
 export * from './useDeviceSize';
-export * from './useDeviceSize2';
 export * from './useCheckImageExists';
 export * from './useRandomColor';
 export * from './usePagination';
