+++ src/hooks/useDeviceSize.ts
@@ -1,14 +1,46 @@
+/* eslint-disable sort-keys-fix/sort-keys-fix */
+/* eslint-disable sort-keys */
+// Here are the dimensions according to the grid SCAN.2.0
 import { useEffect, useMemo, useState } from 'react';
 
 export enum DeviceSize {
-  xxs, // < 567
-  xs, // 767 - 568
-  sm, // 1023 - 768
-  md, // 1279 - 1024
-  lg, // 1439 - 1280
-  xl, // 1919 - 1440
-  xxl // > 1920
+  xxs, // < 480
+  xs, // 575 - 480
+  sm, // 767 - 576
+  md, // 991 - 768
+  lg, // 1199 - 992
+  xl, // 1679 - 1200
+  xxl // > 1680
 }
+
+// for css @media queries
+export const deviceWidth = {
+  only: {
+    xxs: '(max-width: 479px)',
+    xs: '(min-width: 480px) and (max-width: 575px)',
+    sm: '(min-width: 576px) and (max-width: 767px)',
+    md: '(min-width: 768px) and (max-width: 991px)',
+    lg: '(min-width: 992px) and (max-width: 1199px)',
+    xl: '(min-width: 1200px) and (max-width: 1679px)',
+    xxl: '(min-width: 1680px)'
+  },
+  biggerThan: {
+    xxs: '(min-width: 480px)',
+    xs: '(min-width: 576px)',
+    sm: '(min-width: 768px)',
+    md: '(min-width: 992px)',
+    lg: '(min-width: 1200px)',
+    xl: '(min-width: 1680px)'
+  },
+  smallerThan: {
+    xs: '(max-width: 479px)',
+    sm: '(max-width: 575px)',
+    md: '(max-width: 767px)',
+    lg: '(max-width: 991px)',
+    xl: '(max-width: 1199px)',
+    xxl: '(max-width: 1679px)'
+  }
+};
 
 export const useDeviceSize = (): DeviceSize => {
   const [windowWidth, setWindowWidth] = useState<number | undefined>();
@@ -30,12 +62,12 @@ export const useDeviceSize = (): DeviceSize => {
   }, []); // Empty array ensures that effect is only run on mount
 
   return useMemo(() => {
-    if (windowWidth && windowWidth < 567) return DeviceSize.xxs;
-    if (windowWidth && windowWidth < 768) return DeviceSize.xs;
-    if (windowWidth && windowWidth < 1024) return DeviceSize.sm;
-    if (windowWidth && windowWidth < 1280) return DeviceSize.md;
-    if (windowWidth && windowWidth < 1440) return DeviceSize.lg;
-    if (windowWidth && windowWidth < 1920) return DeviceSize.xl;
+    if (windowWidth && windowWidth < 480) return DeviceSize.xxs;
+    if (windowWidth && windowWidth < 576) return DeviceSize.xs;
+    if (windowWidth && windowWidth < 768) return DeviceSize.sm;
+    if (windowWidth && windowWidth < 992) return DeviceSize.md;
+    if (windowWidth && windowWidth < 1200) return DeviceSize.lg;
+    if (windowWidth && windowWidth < 1680) return DeviceSize.xl;
 
     return DeviceSize.xxl;
   }, [windowWidth]);
